Added hotcold flag to db_bench
Summary: In order to benchmark the overhead of metrics taking I added a flag to db_bench that calls `DB::OpenWithHotCold()` instead of `DB::Open()` when it is set.

Test Plan: Added print statements to lines before `DB::Open()` and `DB::OpenWithHotCold()` to see whether the flag affects which one gets called.

Reviewers: vamsi, dhruba

Reviewed By: dhruba

CC: leveldb

Differential Revision: https://reviews.facebook.net/D7485