Hotcold: Short circuiting behaviour for file lookups when finding key in hot file
Summary:
Before this update both hot and cold file had to be queried in order to get the latest value for a given key. Now if the key is found in the hot file then we know the latest value is in the hot file.

Changed compaction to not allow a user key to have a record with a smaller sequence number in the hot file if there is a record with a larger sequence number in the cold file.

Added short circuiting to `Version::Get()` to assume that if a key is found in a newer file then it has a newer sequence number.

Test Plan:
`make check`

ran `db_bench` compiled with assertions on and in hot cold mode.

Reviewers: vamsi, dhruba

Reviewed By: vamsi

CC: leveldb

Differential Revision: https://reviews.facebook.net/D8451