Improved readhotwriterandom db_bench benchmark
Summary: readhotwriterandom now generates a certain number of keys randomly. This number is a percentage (determined by hotkeyspercentage) of --num. Then when a key is needed one of the keys in that list is chosen randomly.

Test Plan: Ran db_bench with small --num parameter to see if the keys are random and that we only use a subset of the "hot" keys.

Reviewers: vamsi, dhruba

Reviewed By: vamsi

CC: leveldb

Differential Revision: https://reviews.facebook.net/D8613